improved responsiveness of Contact page
