+++ frontend/src/ContactForm/ContactForm.js
@@ -10,7 +10,7 @@ const ContactForm = (props) => {
       EmailAddress: "",
       Message: "",
       PhoneNumber1: "",
-      PhoneNumber2: "", 
+      PhoneNumber2: "",
       PhoneNumber3: ""
     })
 
@@ -25,7 +25,7 @@ const ContactForm = (props) => {
       Country: ""
     })
 
-  const handleChangeBasicFormFields = (event) => {  
+  const handleChangeBasicFormFields = (event) => {
     let newBasicFormFields = { ...basicFormFields }
     newBasicFormFields[event.target.name] = event.target.value
     setBasicFormFields(newBasicFormFields)
@@ -64,26 +64,11 @@ const ContactForm = (props) => {
       StateCounty: addressDetails.StateCounty,
       Country: addressDetails.Country
     }
-    // const formLabels =
-    // {
-    //   FullName: "Full name",
-    //   EmailAddress: "Email address",
-    //   PhoneNumbers: "Phone number",
-    //   Message: "Message",
-    //   bIncludeAddressDetails: "Add address details",
-    //   "AddressDetails.AddressLine1": "Address line 1",
-    //   "AddressDetails.AddressLine2": "Address line 2",
-    //   "AddressDetails.CityTown": "City/Town",
-    //   "AddressDetails.StateCounty": "State/County",
-    //   "AddressDetails.Postcode": "Postcode",
-    //   "AddressDetails.Country": "Country"
-    // }
     fetch("http://0.0.0:8080/contact-us", {
       method: "POST",
       body: JSON.stringify(formAnswers),
       headers: {
-         "Content-Type": 'application/json', 
-      //   "Access-Control-Allow-origin": "*"
+        "Content-Type": 'application/json',
       }
     }).then(response => {
       console.log(response)
@@ -94,8 +79,8 @@ const ContactForm = (props) => {
       } else {
         props.setFormSubmitted(false)
         alert(responseBody.Message)
-      } 
-       
+      }
+
     })
   }
 
@@ -108,27 +93,28 @@ const ContactForm = (props) => {
         </div>
         <div className="form_field_and_label">
           <label htmlFor="email">Email address</label>
-          <input type="text" id="email" name="EmailAddress" value={basicFormFields.EmailAddress || ""} required onChange={event => handleChangeBasicFormFields(event)} />
+          <input type="email" id="email" name="EmailAddress" value={basicFormFields.EmailAddress || ""} required onChange={event => handleChangeBasicFormFields(event)} />
         </div>
       </div>
 
       <div className="phone_numbers_section">
-          <div className="form_field_and_label">
-            <label htmlFor="phone_number_1">Phone number 01 <span> - with prefix (e.g. +44123456789) </span></label>
-            <input type="tel" name="PhoneNumber1" id="phone_number_1" value={basicFormFields.PhoneNumber1} onChange={event => handleChangeBasicFormFields(event)} />
-          </div>
+        <div className="form_field_and_label">
+          <label htmlFor="phone_number_1">Phone number 01 <span> - with prefix (e.g. +44123456789) </span></label>
+          <input type="tel" name="PhoneNumber1" pattern="\+[0-9]{1,3}[0-9]{6,14}" required
+            id="phone_number_1" value={basicFormFields.PhoneNumber1} onChange={event => handleChangeBasicFormFields(event)} />
+        </div>
 
-          <div className="form_field_and_label">
-            <label htmlFor="phone_number_2">Phone number 02 <span> - optional </span></label>
-            <input type="tel" name="PhoneNumber2" id="phone_number_2" value={basicFormFields.PhoneNumber2} onChange={event => handleChangeBasicFormFields(event)} />
-          </div>
+        <div className="form_field_and_label">
+          <label htmlFor="phone_number_2">Phone number 02 <span> - optional </span></label>
+          <input type="tel" name="PhoneNumber2" id="phone_number_2" pattern="\+[0-9]{1,3}[0-9]{6,14}" value={basicFormFields.PhoneNumber2} onChange={event => handleChangeBasicFormFields(event)} />
+        </div>
 
-          <div className="form_field_and_label">
-            <label htmlFor="phone_number_3">Phone number 03 <span> - optional </span></label>
-            <input type="tel" name="PhoneNumber3" id="phone_number_3" value={basicFormFields.PhoneNumber3} onChange={event => handleChangeBasicFormFields(event)} />
-          </div>
+        <div className="form_field_and_label">
+          <label htmlFor="phone_number_3">Phone number 03 <span> - optional </span></label>
+          <input type="tel" name="PhoneNumber3" id="phone_number_3" pattern="\+[0-9]{1,3}[0-9]{6,14}" value={basicFormFields.PhoneNumber3} onChange={event => handleChangeBasicFormFields(event)} />
+        </div>
       </div>
-          
+
       <div className="message_section form_field_and_label">
         <label htmlFor="Message">Message</label>
         <textarea name="Message" id="Message" value={basicFormFields.Message || ""} required onChange={event => handleChangeBasicFormFields(event)} />
@@ -152,25 +138,21 @@ const ContactForm = (props) => {
           </div>
 
           <div className="city_and_country_details">
-            <div className="city_county_fields">
-              <div className="form_field_and_label">
-                <label htmlFor="city">City/Town</label>
-                <input type="text" name="CityTown" id="city" value={addressDetails.CityTown || ""} required onChange={event => handleChangeAddress(event)} />
-              </div>
-              <div className="form_field_and_label">
-                <label htmlFor="state">State/County</label>
-                <input type="text" name="StateCounty" id="state" value={addressDetails.StateCounty || ""} required onChange={event => handleChangeAddress(event)} />
-              </div>
+            <div className="form_field_and_label">
+              <label htmlFor="city">City/Town</label>
+              <input type="text" name="CityTown" id="city" value={addressDetails.CityTown || ""} required onChange={event => handleChangeAddress(event)} />
             </div>
-            <div className="postcode_country_fields">
-              <div className="form_field_and_label">
-                <label htmlFor="postcode">Postcode</label>
-                <input type="text" name="Postcode" id="postcode" value={addressDetails.Postcode || ""} required onChange={event => handleChangeAddress(event)} />
-              </div>
-              <div className="form_field_and_label">
-                <label htmlFor="country">Country</label>
-                <input type="text" name="Country" id="country" value={addressDetails.Country || ""} required onChange={event => handleChangeAddress(event)} />
-              </div>
+            <div className="form_field_and_label">
+              <label htmlFor="state">State/County</label>
+              <input type="text" name="StateCounty" id="state" value={addressDetails.StateCounty || ""} onChange={event => handleChangeAddress(event)} />
+            </div>
+            <div className="form_field_and_label">
+              <label htmlFor="postcode">Postcode</label>
+              <input type="text" name="Postcode" id="postcode" value={addressDetails.Postcode || ""} required onChange={event => handleChangeAddress(event)} />
+            </div>
+            <div className="form_field_and_label">
+              <label htmlFor="country">Country</label>
+              <input type="text" name="Country" id="country" value={addressDetails.Country || ""} required onChange={event => handleChangeAddress(event)} />
             </div>
           </div>
         </div>
