+++ frontend/src/ContactForm/ContactForm.scss
@@ -42,15 +42,15 @@ textarea {
     flex-wrap: wrap;
     justify-content: space-between;
     gap: 10px;
-    
-}
 
+}
 
 .phone_numbers_section {
     display: flex;
     flex-direction: column;
     justify-content: space-evenly;
     gap: 10px;
+
     span {
         color: $secondary_text_colour;
         font-size: $paragraph_font_size;
@@ -90,16 +90,11 @@ textarea {
     flex-wrap: wrap;
     justify-content: space-between;
 
-    .city_county_fields, .postcode_country_fields {
-        display: flex;
-        flex-direction: row;
-        justify-content: space-between;
-        gap: 10px;
-    }
     .form_field_and_label {
-        max-width: 100px;
+        max-width: 22%;
     }
 }
+
 .submit_button {
     display: flex;
     flex-direction: row;
@@ -121,14 +116,21 @@ textarea {
 }
 
 @media screen and (max-width: 800px) {
-  
+
+    form {
+        max-width: 100%;
+    }
+
     .city_and_country_details {
-        display:flex;
-        flex-direction: column;
+        display: flex;
         align-items: center;
         justify-content: space-between;
-        gap: 25px;
-   
+        gap: 10px;
+
+        .form_field_and_label {
+            width: 48%;
+        }
+
     }
 
     .name_and_email_section {
