+++ frontend/src/Footer/Footer.js
@@ -3,7 +3,7 @@ const Footer = () => {
     return (
         <footer>
             <p className="main_credits">Website by <a className="main_credits" href="/https://melody-sylvestre.github.io/">Melody Sylvestre</a></p>
-            <p className="additional_credits">Mountain photo by <a className="additional_credits" href="https://unsplash.com/@the_bracketeer">Hendrik Cornelissen</a> and lake photo <a className="additional_credits" href="https://unsplash.com/@luckybeanz">Mark Harpur</a></p>
+            <p className="additional_credits">Additional photos by <a className="additional_credits" href="https://unsplash.com/@the_bracketeer">Hendrik Cornelissen</a> and <a className="additional_credits" href="https://unsplash.com/@luckybeanz">Mark Harpur</a></p>
         </footer>
     )
 }
