+++ frontend/src/Footer/Footer.scss
@@ -15,6 +15,7 @@ footer {
     justify-content: space-between;
     margin-left: $left_margin_desktop;
     margin-right: $right_margin_desktop;
+    margin-top: 20px;
     min-width: auto;
     padding-top: 10px;
